Repositioning of Model class
Repositioning of Model class
Changes to NavBar for dashboard
