+++ Common/Models/CommonAmenities.cs
@@ -4,8 +4,16 @@ namespace SocietyManagementShowcase.Models
 {
     public class CommonAmenities
     {
+        public int Id { get; set; }
         public StatusHealth Health { get; set; }
         public List<MaintenanceLog> CommonAmenitiesMaintenaceLog { get; set; }
         public DateOnly LastMaintenanceCheck { get; set; }
+        public CommonAmenitiesType Type { get; set; }
+    }
+
+    public enum CommonAmenitiesType
+    {
+        Men = 0,
+        Women
     }
 }
