+++ Common/Models/Gym.cs
@@ -4,6 +4,7 @@ namespace SocietyManagementShowcase.Models
 {
     public class Gym
     {
+        public int Id { get; set; }
         public StatusHealth Health { get; set; }
         public List<MaintenanceLog> SwimmingPoolMaintenaceLog { get; set; }
         public DateOnly LastMaintenanceCheck { get; set; }
