+++ Common/Models/Society.cs
@@ -8,16 +8,12 @@ namespace SocietyManagementShowcase.Models
         public string Name { get; set; }
         public List<Wing> Wing { get; set; }
         public int NoOfWings { get; set; }
-        public Person Manager { get; set; }
-        public List<Person> SecurityStaff { get; set; }
-        public List<Person> HouseKeepingStaff { get; set; }
+        public List<Person> Staff { get; set; }
         public FireFightingSystem FireSystem { get; set; }
         public WaterFiltrationSystem WaterFilter { get; set; }
-        public SwimmingPool OutdoorSwimmingPool { get; set; }
-        public SwimmingPool IndoorSwimmingPool { get; set; }
+        public ICollection<SwimmingPool> SwimmingPools { get; set; }        
         public Gym IndoorGym { get; set; }
-        public CommonAmenities CommonAmenitiesMen { get; set; }
-        public CommonAmenities CommonAmenitiesWomen { get; set; }
+        public ICollection<CommonAmenities> CommonAmenities { get; set; }        
         public StatusHealth Health { get; set; }
         public double TotalFlatAreaSociety { get; set; }
         public double TotalMaintenanceChargeSociety { get; set; }
