+++ Common/Models/SwimmingPool.cs
@@ -4,8 +4,16 @@ namespace SocietyManagementShowcase.Models
 {
     public class SwimmingPool
     {
+        public int Id { get; set; }
         public StatusHealth Health { get; set; }
         public List<MaintenanceLog> SwimmingPoolMaintenaceLog { get; set; }
         public DateOnly LastMaintenanceCheck { get; set; }
+        public SwimmingPoolType PoolType { get; set; }
+    }
+
+    public enum SwimmingPoolType
+    {
+        Indoor = 0,
+        Outdoor
     }
 }
