+++ Common/Models/User.cs
@@ -1,15 +1,10 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace SocietyManagementShowcase.Models
+﻿namespace SocietyManagementShowcase.Models
 {
     public class User
     {
         public int Id { get; set; }
-        [Required]
         public string Username { get; set; }
-        [Required]
         public string Password { get; set; }
-        [Required]
         public int PersonId { get; set; }
     }
 }
