--- SocietyManagementShowcase/Models/FireFightingSystem.cs
@@ -1,13 +1,0 @@
-﻿using SocietyManagementShowcase.Common;
-
-namespace SocietyManagementShowcase.Models
-{
-    public class FireFightingSystem
-    {
-        public int Id { get; set; }
-        public WaterTank Fire { get; set; }
-        public Engine FireSystemEngine { get; set; }
-
-        public StatusHealth Health { get; set; }
-    }
-}
