--- SocietyManagementShowcase/Models/Gym.cs
@@ -1,12 +1,0 @@
-﻿using SocietyManagementShowcase.Common;
-
-namespace SocietyManagementShowcase.Models
-{
-    public class Gym
-    {
-        public int Id { get; set; }
-        public StatusHealth Health { get; set; }
-        public List<MaintenanceLog> SwimmingPoolMaintenaceLog { get; set; }
-        public DateOnly LastMaintenanceCheck { get; set; }
-    }
-}
