--- SocietyManagementShowcase/Models/User.cs
@@ -1,10 +1,0 @@
-﻿namespace SocietyManagementShowcase.Models
-{
-    public class User
-    {
-        public int Id { get; set; }
-        public string Username { get; set; }
-        public string Password { get; set; }
-        public int PersonId { get; set; }
-    }
-}
