--- SocietyManagementShowcase/Models/VisitorLog.cs
@@ -1,13 +1,0 @@
-﻿namespace SocietyManagementShowcase.Models
-{
-    public class VisitorLog
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-        public string Address { get; set; }
-        public string Purpose { get; set; }
-        public int PhoneNumber { get; set; }
-        public string Wing { get; set; }
-        public string Flat { get; set; }
-    }
-}
